Create streams.js
Practicing streams with Node
Solution 1 will load the entire file at once. It will work on my local machine but not at scale

Solution 2. This will stream the contents of the data file but we may have an issue called backpressure where the receiving end of the data transfer cannot handle the amount of incoming data

Solution 3 uses a pipe as a workaround, but there may be customized solutions that require the other solutions but generally Solution 3 can be used